pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-shape-generator-mcp"
version = "0.1.0"
description = "MCP Server for AI-powered shape generation in canvas editor"
requires-python = ">=3.9"
dependencies = [
    "mcp>=0.1.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
